Use Jenkinsfile from archetype
Use the Jenkinsfile from the plugin archetype in order to reduce
differences throughout the Jenkins organization.

This intentionally reduces CI costs by running tests in fewer
configurations.  It intentionally skips tests on Java 11 because we are
compiling to Java 11 byte code and testing on Java 21 (Linux) and Java
17 (Windows) with the Java 11 byte code.  It intentionally removes the
Jenkins version specific tests because after they have passed once,
they almost never fail in later tests.

The Jenkinsfile from the plugin archetype is extended to run tests in
parallel in order to reduce the time and cost of tests.

https://github.com/jenkinsci/archetypes/issues/650 describes the
alternatives for Jenkinsfile configurations.