Require Jenkins 2.426.3 or newer
As of 19 Apr 2024, 86% of the installations of the most recent release
(336.v788e4055508b_ - released 3 months ago) are using Jenkins 2.426.3
or newer.  64% of all installations of the plugin are using the most
recent release, 336.v788e4055508b_.  Users that are upgrading the plugin
are already using Jenkins 2.426.3.

Jenkins 2.426.3 is the first version with the fix for
https://www.jenkins.io/security/advisory/2024-01-24/#SECURITY-3314, the
arbitrary file read vulnerability through the CLI can lead to RCE. It
is a very good choice as a minimum Jenkins version.

Jenkins 2.426.3 is one of the versions suggested by
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

Updates the plugin bill of matertials for Jenkins 2.426.x as well