Require Jenkins 2.440.3 or newer (#130)
https://stats.jenkins.io/pluginversions/cloud-stats.html shows that
Jenkins 2.440.3 or newer is already used by 80% of the installations
of the most recent release and the most recent release is used in 75%
of all installations.  Jenkins 2.440.3 also includes a security fix that
is good for users to deploy.

BOM updates have stopped for bom-2.426.x so this will also continue
updating plugin dependencies by switching to a bom version that is still
receiving updates.

Change also removes the duplication of the Jenkins baseline as has been
done recently in the plugin archetype.  Refer to:

* https://github.com/jenkinsci/archetypes/pull/737